Move sanitize_raw_request to image.rs
