+++ tensorzero-internal/src/endpoints/inference.rs
@@ -719,13 +719,15 @@ async fn write_inference(
         // Write the inference to the Inference table
         match result {
             InferenceResult::Chat(result) => {
-                let chat_inference = ChatInferenceDatabaseInsert::new(result, input.clone(), metadata);
+                let chat_inference =
+                    ChatInferenceDatabaseInsert::new(result, input.clone(), metadata);
                 let _ = clickhouse_connection_info
                     .write(&[chat_inference], "ChatInference")
                     .await;
             }
             InferenceResult::Json(result) => {
-                let json_inference = JsonInferenceDatabaseInsert::new(result, input.clone(), metadata);
+                let json_inference =
+                    JsonInferenceDatabaseInsert::new(result, input.clone(), metadata);
                 let _ = clickhouse_connection_info
                     .write(&[json_inference], "JsonInference")
                     .await;
